scripts/SD/pyproject.toml
[tool.ruff]
line-length = 500
[tool.poetry]
name = "ai-horde"
version = "0.1.0"
description = ""
authors = ["AmdaUwU <jbat2000@gmail.com>"]
readme = "README.md"
packages = [{include = "ai_horde"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = ">=2.27"
loguru = ">=0.6.0,<0.7.0"
pillow = ">=9.3.0,<9.4.0"
werkzeug = ">=2.2.2,<2.3.0"
flask = ">=2.2.2,<2.3.0"
flask-restx = "^1.1.0"
flask-limiter = ">=2.8.1,<2.9.0"
flask-caching = "^2.0.2"
waitress = ">=2.1.2,<2.2.0"
markdown = ">=3.4.1,<3.5.0"
flask-dance = {extras = ["sqla"], version = "^6.2.0"}
blinker = "^1.5"
python-dotenv = "^1.0.0"
bleach = ">=5.0.1,<5.1.0"
alt-profanity-check = "^1.2.1"
better-profanity = "^0.7.0"
python-dateutil = ">=2.8.2,<2.9.0"
redis = ">=4.3.5,<4.4.0"
flask-sqlalchemy = "3.0.2"
oauthlib = ">=3.2.2,<3.3.0"
sqlalchemy = ">=1.4.44,<1.5.0"
psycopg2-binary = "^2.9.5"
boto3 = "^1.26.85"
regex = "^2022.10.31"
unidecode = "^1.3.6"
patreon = {git = "https://github.com/Patreon/patreon-python.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
